+++ src/app/(providers)/(root)/tour/[id]/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const page = () => {
-  return <div>page</div>;
+  return <div></div>;
 };
 
 export default page;
