+++ src/app/(providers)/(root)/tour/page.tsx
@@ -1,10 +1,13 @@
 import React from 'react';
+import TourList from '@/components/tour/TourList';
 
 const TourPage = () => {
   return (
     <div>
       Travel Package
-      <div></div>
+      <div>
+        <TourList />
+      </div>
     </div>
   );
 };
